# Conflicts:
#	CHANGELOG.md
#	README.md
#	dependencies.list
#	examples/architectureComponentsExample/build.gradle
#	examples/multiprocessExample/build.gradle
#	realm/realm-library/build.gradle
#	realm/realm-library/src/main/cpp/io_realm_SyncSession.cpp
#	realm/realm-library/src/main/cpp/io_realm_internal_OsObject.cpp
#	realm/realm-library/src/main/cpp/object-store
#	realm/realm-library/src/objectServer/java/io/realm/SyncConfiguration.java
#	tools/sync_test_server/Dockerfile
#	version.txt

+++ realm/realm-library/src/androidTest/java/io/realm/entities/IOSAllTypes.java
@@ -32,7 +32,7 @@ public class IOSAllTypes extends RealmObject {
     private boolean boolCol;
     private short shortCol;
     private int intCol;
-    private int longCol;
+    private long longCol;
     private long longLongCol;
     private float floatCol;
     private double doubleCol;
@@ -74,11 +74,11 @@ public class IOSAllTypes extends RealmObject {
         this.intCol = intCol;
     }
 
-    public int getLongCol() {
+    public long getLongCol() {
         return longCol;
     }
 
-    public void setLongCol(int longCol) {
+    public void setLongCol(long longCol) {
         this.longCol = longCol;
     }
